Fixed issue with comma, new lines & the jpaDerivedIdentifier option
When having a relationship like this:
```
relationship OneToOne {
  A to B with jpaDerivedIdentifier,
  C to D
}
```

The parsing system didn't know if the comma was separating the relationship
from A to B or the relationship option list.

The comma is supposed to separate both, but a restriction has been made in this
commit: a comma without a new line has to separate relationship options.

Fixes #406